+++ src/egon/data/airflow/dags/pipeline.py
@@ -5,35 +5,38 @@ from airflow.operators.python_operator import PythonOperator
 from airflow.utils.dates import days_ago
 import importlib_resources as resources
 
-from egon.data.airflow.tasks import initdb
+from egon.data.datasets import database
+from egon.data.datasets.osm import OpenStreetMap
 from egon.data.processing.zensus_vg250 import (
     zensus_population_inside_germany as zensus_vg250,
 )
 import airflow
 import egon.data.importing.demandregio as import_dr
+import egon.data.importing.demandregio.install_disaggregator as install_dr
+import egon.data.importing.era5 as import_era5
 import egon.data.importing.etrago as etrago
 import egon.data.importing.heat_demand_data as import_hd
+import egon.data.importing.industrial_sites as industrial_sites
 import egon.data.importing.mastr as mastr
-import egon.data.importing.openstreetmap as import_osm
+import egon.data.importing.nep_input_data as nep_input
 import egon.data.importing.re_potential_areas as re_potential_areas
+import egon.data.importing.scenarios as import_scenarios
 import egon.data.importing.vg250 as import_vg250
-import egon.data.processing.demandregio as process_dr
-import egon.data.processing.openstreetmap as process_osm
 import egon.data.importing.zensus as import_zs
-import egon.data.processing.zensus as process_zs
+import egon.data.processing.boundaries_grid_districts as boundaries_grid_districts
+import egon.data.processing.demandregio as process_dr
+import egon.data.processing.district_heating_areas as district_heating_areas
+import egon.data.processing.loadarea as loadarea
 import egon.data.processing.osmtgmod as osmtgmod
 import egon.data.processing.power_plants as power_plants
-import egon.data.importing.nep_input_data as nep_input
-import egon.data.importing.etrago as etrago
-import egon.data.importing.mastr as mastr
+import egon.data.processing.renewable_feedin as import_feedin
 import egon.data.processing.substation as substation
 import egon.data.processing.zensus_vg250.zensus_population_inside_germany as zensus_vg250
-import egon.data.importing.re_potential_areas as re_potential_areas
-import egon.data.importing.heat_demand_data as import_hd
-import egon.data.importing.scenarios as import_scenarios
 import egon.data.importing.gas_grid as gas_grid
+import egon.data.processing.mv_grid_districts as mvgd
+import egon.data.processing.zensus as process_zs
+import egon.data.processing.zensus_grid_districts as zensus_grid_districts
 
-import egon.data.importing.industrial_sites as industrial_sites
 from egon.data import db
 
 
@@ -51,26 +54,17 @@ with airflow.DAG(
     is_paused_upon_creation=False,
     schedule_interval=None,
 ) as pipeline:
-    setup = PythonOperator(task_id="initdb", python_callable=initdb)
 
-    # Openstreetmap data import
-    osm_download = PythonOperator(
-        task_id="download-osm",
-        python_callable=import_osm.download_pbf_file,
-    )
-    osm_import = PythonOperator(
-        task_id="import-osm",
-        python_callable=import_osm.to_postgres,
-    )
-    osm_migrate = PythonOperator(
-        task_id="migrate-osm",
-        python_callable=process_osm.modify_tables,
-    )
-    osm_add_metadata = PythonOperator(
-        task_id="add-osm-metadata",
-        python_callable=import_osm.add_metadata,
-    )
-    setup >> osm_download >> osm_import >> osm_migrate >> osm_add_metadata
+    tasks = pipeline.task_dict
+
+    database_setup = database.Setup()
+    database_setup.insert_into(pipeline)
+    setup = tasks["database.setup"]
+
+    osm = OpenStreetMap(dependencies=[setup])
+    osm.insert_into(pipeline)
+    osm_add_metadata = tasks["osm.add-metadata"]
+    osm_download = tasks["osm.download"]
 
     # VG250 (Verwaltungsgebiete 250) data import
     vg250_download = PythonOperator(
@@ -182,10 +176,20 @@ with airflow.DAG(
 
     scenario_input_tables >> demandregio_tables
 
+
+    demandregio_installation = PythonOperator(
+        task_id="demandregio-installation",
+        python_callable=install_dr.clone_and_install,
+    )
+
+    setup >> demandregio_installation
+
     demandregio_society = PythonOperator(
         task_id="demandregio-society",
         python_callable=import_dr.insert_society_data,
     )
+
+    demandregio_installation >> demandregio_society
     vg250_clean_and_prepare >> demandregio_society
     demandregio_tables >> demandregio_society
     scenario_input_import >> demandregio_society
@@ -194,6 +198,8 @@ with airflow.DAG(
         task_id="demandregio-household-demands",
         python_callable=import_dr.insert_household_demand,
     )
+
+    demandregio_installation >> demandregio_demand_households
     vg250_clean_and_prepare >> demandregio_demand_households
     demandregio_tables >> demandregio_demand_households
     scenario_input_import >> demandregio_demand_households
@@ -202,6 +208,8 @@ with airflow.DAG(
         task_id="demandregio-cts-industry-demands",
         python_callable=import_dr.insert_cts_ind_demands,
     )
+
+    demandregio_installation >> demandregio_demand_cts_ind
     vg250_clean_and_prepare >> demandregio_demand_cts_ind
     demandregio_tables >> demandregio_demand_cts_ind
     scenario_input_import >> demandregio_demand_cts_ind
@@ -233,6 +241,7 @@ with airflow.DAG(
     demandregio_society >> household_prognosis
     zensus_misc_import >> household_prognosis
 
+
     # Distribute electrical demands to zensus cells
     processed_dr_tables = PythonOperator(
         task_id="create-demand-tables",
@@ -248,13 +257,6 @@ with airflow.DAG(
     population_prognosis >> elec_household_demands_zensus
     demandregio_demand_households >> elec_household_demands_zensus
     map_zensus_vg250 >> elec_household_demands_zensus
-
-    # Power plant setup
-    power_plant_tables = PythonOperator(
-        task_id="create-power-plant-tables",
-        python_callable=power_plants.create_tables,
-    )
-    setup >> power_plant_tables
 
     # NEP data import
     create_tables = PythonOperator(
@@ -310,16 +312,18 @@ with airflow.DAG(
         autocommit=True,
     )
 
-    create_voronoi = PythonOperator(
-        task_id="create_voronoi", python_callable=substation.create_voronoi
-    )
     osm_add_metadata >> substation_tables >> substation_functions
-    substation_functions >> hvmv_substation_extraction >> create_voronoi
-    substation_functions >> ehv_substation_extraction >> create_voronoi
+    substation_functions >> hvmv_substation_extraction
+    substation_functions >> ehv_substation_extraction
     vg250_clean_and_prepare >> hvmv_substation_extraction
     vg250_clean_and_prepare >> ehv_substation_extraction
 
     # osmTGmod ehv/hv grid model generation
+    osmtgmod_osm_import = PythonOperator(
+        task_id="osmtgmod_osm_import",
+        python_callable=osmtgmod.import_osm_data,
+    )
+
     run_osmtgmod = PythonOperator(
         task_id="run_osmtgmod",
         python_callable=osmtgmod.run_osmtgmod,
@@ -330,9 +334,33 @@ with airflow.DAG(
         python_callable=osmtgmod.osmtgmmod_to_pypsa,
     )
 
+    osmtgmod_substation = PostgresOperator(
+        task_id="osmtgmod_substation",
+        sql=resources.read_text(osmtgmod, "substation_otg.sql"),
+        postgres_conn_id="egon_data",
+        autocommit=True,
+    )
+
+    osm_download >> osmtgmod_osm_import >> run_osmtgmod
     ehv_substation_extraction >> run_osmtgmod
     hvmv_substation_extraction >> run_osmtgmod
     run_osmtgmod >> osmtgmod_pypsa
+    etrago_input_data >> osmtgmod_pypsa
+    run_osmtgmod >> osmtgmod_substation
+
+    # MV grid districts
+    create_voronoi = PythonOperator(
+        task_id="create_voronoi",
+        python_callable=substation.create_voronoi
+    )
+    osmtgmod_substation >> create_voronoi
+
+
+    define_mv_grid_districts = PythonOperator(
+        task_id="define_mv_grid_districts",
+        python_callable=mvgd.define_mv_grid_districts
+    )
+    create_voronoi >> define_mv_grid_districts
 
     # Import potential areas for wind onshore and ground-mounted PV
     download_re_potential_areas = PythonOperator(
@@ -366,12 +394,14 @@ with airflow.DAG(
     )
 
     power_plant_import = PythonOperator(
-        task_id="import-power-plants",
+        task_id="import-hydro-biomass-power-plants",
         python_callable=power_plants.insert_power_plants,
     )
+
     setup >> power_plant_tables >> power_plant_import
     nep_insert_data >> power_plant_import
     retrieve_mastr_data >> power_plant_import
+    define_mv_grid_districts >> power_plant_import
 
     # Import and merge data on industrial sites from different sources
 
@@ -404,6 +434,7 @@ with airflow.DAG(
     demandregio_demand_cts_ind >> elec_cts_demands_zensus
     map_zensus_vg250 >> elec_cts_demands_zensus
 
+
     # Gas grid import
     gas_grid_insert_data = PythonOperator(
         task_id="insert-gas-grid",
@@ -411,3 +442,115 @@ with airflow.DAG(
     )
 
     create_tables >> gas_grid_insert_data
+
+    # Extract landuse areas from osm data set
+    create_landuse_table = PythonOperator(
+        task_id="create-landuse-table",
+        python_callable=loadarea.create_landuse_table
+    )
+
+    landuse_extraction = PostgresOperator(
+        task_id="extract-osm_landuse",
+        sql=resources.read_text(loadarea, "osm_landuse_extraction.sql"),
+        postgres_conn_id="egon_data",
+        autocommit=True,
+    )
+    setup >> create_landuse_table
+    create_landuse_table >> landuse_extraction
+    osm_add_metadata >> landuse_extraction
+    vg250_clean_and_prepare >> landuse_extraction
+
+ # Import weather data
+    download_era5 = PythonOperator(
+        task_id="download-weather-data",
+        python_callable=import_era5.download_era5,
+    )
+    scenario_input_import >> download_era5
+
+    create_weather_tables = PythonOperator(
+        task_id="create-weather-tables",
+        python_callable=import_era5.create_tables,
+    )
+    setup >> create_weather_tables
+
+    import_weather_cells = PythonOperator(
+        task_id="insert-weather-cells",
+        python_callable=import_era5.insert_weather_cells,
+    )
+    create_weather_tables >> import_weather_cells
+    download_era5 >> import_weather_cells
+
+    feedin_wind_onshore = PythonOperator(
+        task_id="insert-feedin-wind",
+        python_callable=import_feedin.wind_feedin_per_weather_cell,
+    )
+
+    feedin_pv = PythonOperator(
+        task_id="insert-feedin-pv",
+        python_callable=import_feedin.pv_feedin_per_weather_cell,
+    )
+
+    feedin_solar_thermal = PythonOperator(
+        task_id="insert-feedin-solar-thermal",
+        python_callable=import_feedin.solar_thermal_feedin_per_weather_cell,
+    )
+
+    import_weather_cells >> [feedin_wind_onshore,
+                             feedin_pv, feedin_solar_thermal]
+    vg250_clean_and_prepare >> [feedin_wind_onshore,
+                             feedin_pv, feedin_solar_thermal]
+
+    # District heating areas demarcation
+    create_district_heating_areas_table = PythonOperator(
+        task_id="create-district-heating-areas-table",
+        python_callable=district_heating_areas.create_tables
+    )
+    import_district_heating_areas = PythonOperator(
+        task_id="import-district-heating-areas",
+        python_callable=district_heating_areas.
+        district_heating_areas_demarcation
+    )
+    setup >> create_district_heating_areas_table
+    create_district_heating_areas_table >> import_district_heating_areas
+    zensus_misc_import >> import_district_heating_areas
+    heat_demand_import >> import_district_heating_areas
+    scenario_input_import >> import_district_heating_areas
+
+    # Electrical load curves CTS
+    map_zensus_grid_districts = PythonOperator(
+        task_id="map_zensus_grid_districts",
+        python_callable=zensus_grid_districts.map_zensus_mv_grid_districts,
+    )
+    population_import >> map_zensus_grid_districts
+    define_mv_grid_districts >> map_zensus_grid_districts
+
+    electrical_load_curves_cts = PythonOperator(
+        task_id="electrical-load-curves-cts",
+        python_callable=process_dr.insert_cts_load,
+    )
+    map_zensus_grid_districts >> electrical_load_curves_cts
+    elec_cts_demands_zensus >> electrical_load_curves_cts
+    demandregio_demand_cts_ind >> electrical_load_curves_cts
+    map_zensus_vg250 >> electrical_load_curves_cts
+    etrago_input_data >> electrical_load_curves_cts
+
+    # Map federal states to mv_grid_districts
+    map_boundaries_grid_districts = PythonOperator(
+        task_id="map_vg250_grid_districts",
+        python_callable=boundaries_grid_districts.map_mvgriddistricts_vg250,
+    )
+    define_mv_grid_districts >> map_boundaries_grid_districts
+    vg250_clean_and_prepare >> map_boundaries_grid_districts
+
+    # Solar rooftop per mv grid district
+    solar_rooftop_etrago = PythonOperator(
+        task_id="etrago_solar_rooftop",
+        python_callable=power_plants.pv_rooftop_per_mv_grid,
+    )
+    map_boundaries_grid_districts >> solar_rooftop_etrago
+    feedin_pv >> solar_rooftop_etrago
+    elec_cts_demands_zensus >> solar_rooftop_etrago
+    elec_household_demands_zensus >> solar_rooftop_etrago
+    nep_insert_data >> solar_rooftop_etrago
+    etrago_input_data >> solar_rooftop_etrago
+    map_zensus_grid_districts >> solar_rooftop_etrago